Fixes generation of UniformMatrix4fv to match expected function signiture
When passing a 4x4 matrix to a shader through a uniform, "glUniformMatrix4fv" expects a "float[]" as input. However, the current implementation of "GL.UniformMatrix4fv" overrides this and expects a reference to a "float". This makes the function unusable as far as I could tell (and is completely different from the implementation by opengl and the other functions in this library).

In the "Generator" project, there is an override for "glUniformMatrix4fv" to change the value to this "ref float". Removing these two lines allows the generator to generate the correct function parameters.