Enables generation of copies of functions with appended descriptors
Modifies the generator to enable generation of new functions with unique names to link to already used OpenGL functions. This allows creation of functions with parameters that are not identical to what is expected by OpenGL, but still valid.

Currently the solution is hard coded to "ByRef" but could be easily adapted to support a list of possible appended names for other functions.